refactor: moving server serving from lib and using a trait for health check telemetry
