+++ crates/apollo-mcp-server/Cargo.toml
@@ -12,6 +12,7 @@ apollo-compiler.workspace = true
 apollo-federation.workspace = true
 apollo-mcp-registry = { path = "../apollo-mcp-registry" }
 apollo-schema-index = { path = "../apollo-schema-index" }
+async-trait = "0.1.88"
 axum = "0.8.4"
 axum-extra = { version = "0.10.1", features = ["typed-header"] }
 bon = "3.6.3"
