+++ crates/apollo-mcp-server/src/lib.rs
@@ -11,3 +11,7 @@ pub mod operations;
 pub mod sanitize;
 pub(crate) mod schema_tree_shake;
 pub mod server;
+pub mod server_config;
+pub mod telemetry;
+
+pub mod server_handler;
