+++ crates/apollo-mcp-server/src/main.rs
@@ -1,5 +1,4 @@
-use std::path::PathBuf;
-
+use crate::runtime::Serve;
 use apollo_mcp_registry::platform_api::operation_collections::collection_poller::CollectionSource;
 use apollo_mcp_registry::uplink::persisted_queries::ManifestSource;
 use apollo_mcp_registry::uplink::schema::SchemaSource;
@@ -7,11 +6,18 @@ use apollo_mcp_server::custom_scalar_map::CustomScalarMap;
 use apollo_mcp_server::errors::ServerError;
 use apollo_mcp_server::operations::OperationSource;
 use apollo_mcp_server::server::Server;
+use apollo_mcp_server::server_config::ServerConfig;
+use apollo_mcp_server::server_handler::ApolloMcpServerHandler;
+use apollo_mcp_server::telemetry::{InMemoryTelemetry, Telemetry};
 use clap::Parser;
 use clap::builder::Styles;
 use clap::builder::styling::{AnsiColor, Effects};
 use runtime::IdOrDefault;
 use runtime::logging::Logging;
+use std::path::PathBuf;
+use std::sync::Arc;
+use tokio::sync::RwLock;
+use tokio_util::sync::CancellationToken;
 use tracing::{info, warn};
 
 mod runtime;
@@ -109,11 +115,15 @@ async fn main() -> anyhow::Result<()> {
         .then(|| config.graphos.graph_ref())
         .transpose()?;
 
-    Ok(Server::builder()
-        .transport(config.transport)
-        .schema_source(schema_source)
-        .operation_source(operation_source)
-        .endpoint(config.endpoint.into_inner())
+    let telemetry: Option<Arc<dyn Telemetry>> = config
+        .health_check
+        .enabled
+        .then(|| Arc::new(InMemoryTelemetry::new()) as Arc<dyn Telemetry>);
+    let server_handler =
+        ApolloMcpServerHandler::new(config.headers.clone(), config.endpoint.clone(), telemetry);
+    let cancellation_token = CancellationToken::new();
+
+    let server_config = ServerConfig::builder()
         .maybe_explorer_graph_ref(explorer_graph_ref)
         .headers(config.headers)
         .execute_introspection(config.introspection.execute.enabled)
@@ -133,8 +143,25 @@ async fn main() -> anyhow::Result<()> {
         )
         .search_leaf_depth(config.introspection.search.leaf_depth)
         .index_memory_bytes(config.introspection.search.index_memory_bytes)
-        .health_check(config.health_check)
+        .build();
+
+    Server::builder()
+        .schema_source(schema_source)
+        .operation_source(operation_source)
+        .server_handler(Arc::new(RwLock::new(server_handler.clone())))
+        .cancellation_token(cancellation_token.child_token())
+        .server_config(server_config)
         .build()
         .start()
-        .await?)
+        .await?;
+
+    Serve::serve(
+        server_handler,
+        config.transport,
+        cancellation_token,
+        config.health_check,
+    )
+    .await?;
+
+    Ok(())
 }
