+++ crates/apollo-mcp-server/src/server/states/operations_configured.rs
@@ -1,12 +1,11 @@
 use apollo_compiler::{Schema, validation::Valid};
 use tracing::debug;
 
+use crate::server_config::ServerConfig;
 use crate::{errors::ServerError, operations::RawOperation, server::states::Starting};
 
-use super::Config;
-
 pub(super) struct OperationsConfigured {
-    pub(super) config: Config,
+    pub(super) config: ServerConfig,
     pub(super) operations: Vec<RawOperation>,
 }
